+++ src/predDesc.h
@@ -349,7 +349,7 @@ class Pred::Forall : public PredDesc {
             body.getAllVars(accu);
         }
         void substFreshId(const std::string &id, const VarName &v){
-            for(int i=0;i<vars.size();i++){
+            for(size_t i=0;i<vars.size();i++){
                 if(vars[i].name.kind() == VarName::Kind::FreshId && vars[i].name.prefix() == id)
                     vars[i].name = v;
             }
@@ -433,7 +433,7 @@ class Pred::Exists : public PredDesc {
             body.getAllVars(accu);
         }
         void substFreshId(const std::string &id, const VarName &v){
-            for(int i=0;i<vars.size();i++){
+            for(size_t i=0;i<vars.size();i++){
                 if(vars[i].name.kind() == VarName::Kind::FreshId && vars[i].name.prefix() == id)
                     vars[i].name = v;
             }
